Luminance-GS/examples/requirements.txt
# assume torch is already installed

# pycolmap for data parsing
git+https://github.com/rmbrualla/pycolmap@cc7ea4b7301720ac29287dbe450952511b32125e
# (optional) nerfacc for torch version rasterization 
# git+https://github.com/nerfstudio-project/nerfacc

viser
nerfview==0.0.2
imageio[ffmpeg]
numpy==1.24.3
scikit-learn
tqdm
torchmetrics[image]==1.4.0
opencv-python
tyro
Pillow
tensorboard==2.17.0
matplotlib
einops
transformers
kornia
